# perf(config/file): optimize file format func